experimenting with the movement system - mouse movement and bumping up the calculation accuracy, although i might need to dial it back down later

+++ core/units/Units.hpp
@@ -60,12 +60,12 @@ inline Vec2 calcDirVec(const Vec2& src, const Vec2& dst) {
 	return { dst.x - src.x, dst.y - src.y };
 }
 
-inline float calcLength(const Vec2& v) {
-	return sqrt(v.x * v.x + v.y * v.y);
+inline double calcLength(const Vec2& v) {
+	return sqrt(double(v.x) * v.x + double(v.y) * v.y);
 }
 
-inline Vec2 scale(const Vec2& v, float s) {
-	return { v.x * s, v.y * s };
+inline Vec2 scale(const Vec2& v, double s) {
+	return { float(v.x * s), float(v.y * s) };
 }
 
 inline Vec2 addVec(const Vec2& a, const Vec2& b) {
