+++ engine/Movement.cpp
@@ -8,8 +8,8 @@ void Movement::start(int destX, int destY, int movSpeed) {
 	startPos = pos;
 	speed = movSpeed;
 	dir = Core::Units::calcDirVec(pos, {float(destX), float(destY)});
-	const float distance = Core::Units::calcLength(dir);
-	timeToDestMs = int(distance / speed * 1000.0f);
+	const double distance = Core::Units::calcLength(dir);
+	timeToDestMs = distance / speed * 1000.0;
 	clock.reset();
 }
 
@@ -27,7 +27,7 @@ void Movement::update() {
 		return;
 	}
 
-	const float timeStep = timeSinceStartMs / float(timeToDestMs);
+	const double timeStep = timeSinceStartMs / timeToDestMs;
 	const Core::Units::Vec2 dirStep = Core::Units::scale(dir, timeStep);
 	pos = Core::Units::addVec(startPos, dirStep);
 }
